+++ Anvil/ecs/lua_ecs.cpp
@@ -5,7 +5,7 @@
 #include <Sprocket/Scripting/lua_converter.h>
 #include <Sprocket/Scripting/lua_script.h>
 #include <Sprocket/Core/log.h>
-#include <Sprocket/Utility/Maths.h>
+#include <Sprocket/Utility/maths.h>
 
 #include <lua.hpp>
 #include <glm/gtc/matrix_transform.hpp>
