+++ Game/ecs/loader.cpp
@@ -5,7 +5,7 @@
 #include "scene.h"
 
 #include <Sprocket/Core/log.h>
-#include <Sprocket/Utility/Yaml.h>
+#include <Sprocket/Utility/yaml.h>
 
 #include <yaml-cpp/yaml.h>
